remove default_branch debugging
Repo is not set up by default. Probably could have used action/checkout,
but might be wasteful if the tree is large?

Instead switch to `gh api`, which doesn't need a local repo. GH_TOKEN is
required.

The default branch seems to be available via
`github.event.repository.default_branch`, but it might not be available
in all situations. And would it cause an error if it wasn't available?
(do Actions workflow expressions do automatic null coalescing?)